fix 3rd file issue.

+++ ch55xduino/ch55x/libraries/Generic_Examples/examples/05.USB/MassStorage/src/userMassStorage/USBMediumAccess.c
@@ -208,7 +208,7 @@ void LUN_Read_func_Root_DIR(uint16_t rootAddrIndex){    //separate funcs relieve
     }
 }
 
-void LUN_Read_func_Files(uint16_t file_data_index){    //separate funcs relieve the register usage
+void LUN_Read_func_Files(uint32_t file_data_index){    //separate funcs relieve the register usage
     uint8_t fileIndex = file_data_index/32768;
     uint16_t fileOffset = file_data_index%32768;
     uint16_t fileSize = filesOnDrive[fileIndex].filesize;
